Fix bug when encountering overlapping deletions.
Fix wrong update of deletion length when a big deletion overlapped a smaller deletion in another sample.
Fix bug in likelihood calculation of regions with extremely high coverage. This fix also slightly reduces the number of false positives called in those regions.
Set default value of -s option to 0.1, as the previous value of 0.75 was much too strict.